Add build.sh and adapt the code to work on the server.
We have a small wart to fix later: in app.py top, we are configuring
static file path to have /scheduler/static/ if the code is meant
to be deployed on production server. However, when developing locally,
we want the app to just use /static/. We did not put any automatic
handling of this, so this has to be manually set.

In this commit, this defaults to just /static/, which will work for
local development, but if this code is deployed to server with build.sh
without editing app.py at the top, the code at the server will be
broken.